requirements.txt
click==8.1.7
psycopg2-binary==2.9.9
prettytable==3.9.0
pytest==8.2.2